Trekk ut kall til dispatch i egen actioncreator
